Several changes in statistics related to hg stats
Changed table names from `stats_mysql_hostgroups(_reset)` to `stats_mysql_hostgroups(_reset)`
Added spaces in around commas in table definition, required for formatting.
Changed `waiting_in_hg` from `int64_t` to `int` because HG is an `int`.
Renaming `sessions_wait*` in `conns_reqs_wait*`
Removed the update of metrics from `MySQL_Backend::reset()

+++ include/MySQL_Session.h
@@ -247,7 +247,7 @@ class MySQL_Session
 	int pending_connect;
 	enum proxysql_session_type session_type;
 	/* @brief Keeps track in which MyHGCs the session is already registered. */
-	int64_t waiting_in_hg;
+	int waiting_in_hg;
 	uint64_t conn_pull_next_wait_start;
 
 	// bool
